+++ app/models/administrator.rb
@@ -1,9 +1,13 @@
 class Administrator < ApplicationRecord
-  def password=(raw_password)
-    if raw_password.kind_of?(String)
-      self.hashed_password = BCrypt::Password.create(raw_password)
-    elsif raw_password.nil?
-      self.hashed_password = nil
-    end
-  end
+  include EmailHolder
+  include PasswordHolder
+
+  # concernsへ移動
+  # def password=(raw_password)
+  #   if raw_password.kind_of?(String)
+  #     self.hashed_password = BCrypt::Password.create(raw_password)
+  #   elsif raw_password.nil?
+  #     self.hashed_password = nil
+  #   end
+  # end
 end
