+++ app/models/customer.rb
@@ -1,4 +1,8 @@
 class Customer < ApplicationRecord
+  include EmailHolder
+  include PersonalNameHolder
+  include PasswordHolder
+
   has_one :home_address, dependent: :destroy, autosave: true
   has_one :work_address, dependent: :destroy, autosave: true
 
@@ -9,11 +13,12 @@ class Customer < ApplicationRecord
     allow_blank: true
   }
 
-  def password=(raw_password)
-    if raw_password.kind_of?(String)
-      self.hashed_password = BCrypt::Password.create(raw_password)
-    elsif raw_password.nil?
-      self.hashed_password = nil
-    end
-  end
+  # concernsへ移動
+  # def password=(raw_password)
+  #   if raw_password.kind_of?(String)
+  #     self.hashed_password = BCrypt::Password.create(raw_password)
+  #   elsif raw_password.nil?
+  #     self.hashed_password = nil
+  #   end
+  # end
 end
