+++ app/models/staff_member.rb
@@ -1,26 +1,31 @@
 class StaffMember < ApplicationRecord
-  include StringNormalizer
+  include EmailHolder
+  include PersonalNameHolder
+  include PasswordHolder
 
   has_many :events, class_name: "StaffEvent", dependent: :destroy
   # has_many :staff_events, dependent: :destroy　通常の書き方ではあるがメソッドチェーンにした時に冗長になりやすい
 
-  before_validation do
-    self.email = normalize_as_email(email)
-    self.family_name = normalize_as_name(family_name)
-    self.given_name = normalize_as_name(given_name)
-    self.family_name_kana = normalize_as_furigana(family_name_kana)
-    self.given_name_kana = normalize_as_furigana(given_name_kana)
-  end
+  # concernsへ移動
+    # before_validation do
+    #   self.email = normalize_as_email(email)
+    #   self.family_name = normalize_as_name(family_name)
+    #   self.given_name = normalize_as_name(given_name)
+    #   self.family_name_kana = normalize_as_furigana(family_name_kana)
+    #   self.given_name_kana = normalize_as_furigana(given_name_kana)
+    # end
 
-  HUMAN_NAME_REGEXP = /\A[\p{han}\p{hiragana}\p{katakana}\u{30fc}A-Za-z]+\z/
-  KATAKANA_REGEXP = /\A[\p{katakana}\u{30fc}]+\z/
+  # concernsへ移動
+    # HUMAN_NAME_REGEXP = /\A[\p{han}\p{hiragana}\p{katakana}\u{30fc}A-Za-z]+\z/
+    # KATAKANA_REGEXP = /\A[\p{katakana}\u{30fc}]+\z/
 
-  validates :email, presence: true, "valid_email_2/email": true,
-    uniqueness: { case_sensitive: false }
-  validates :family_name, :given_name, presence: true,
-    format: { with: HUMAN_NAME_REGEXP, allow_blank: true }
-  validates :family_name_kana, :given_name_kana, presence: true,
-    format: { with: KATAKANA_REGEXP, allow_blank: true } # withオプションで正規化とフリガナ表現を制限
+  # concernsへ移動
+    # validates :email, presence: true, "valid_email_2/email": true,
+    #   uniqueness: { case_sensitive: false }
+    # validates :family_name, :given_name, presence: true,
+    #   format: { with: HUMAN_NAME_REGEXP, allow_blank: true }
+    # validates :family_name_kana, :given_name_kana, presence: true,
+    #   format: { with: KATAKANA_REGEXP, allow_blank: true } # withオプションで正規化とフリガナ表現を制限
   validates :start_date, presence: true, date: {
     after_or_equal_to: Date.new(2000, 1, 1),
     before: -> (obj) { 1.year.from_now.to_date },
@@ -32,13 +37,14 @@ class StaffMember < ApplicationRecord
     allow_blank: true
   }
 
-  def password=(raw_password)
-    if raw_password.kind_of?(String)
-      self.hashed_password = BCrypt::Password.create(raw_password)
-    elsif raw_password.nil?
-      self.hashed_password = nil
-    end
-  end
+  # concernsへ移動
+  # def password=(raw_password)
+  #   if raw_password.kind_of?(String)
+  #     self.hashed_password = BCrypt::Password.create(raw_password)
+  #   elsif raw_password.nil?
+  #     self.hashed_password = nil
+  #   end
+  # end
 
   def active?
     !suspended? && start_date <= Date.today &&
